requirements.txt
Django==1.11.5
django-ckeditor==5.3.0
django-js-asset==0.1.1
django-suit==0.2.25
html5lib==1.0b8
olefile==0.44
Pillow==4.2.1
PyPDF2==1.26.0
pytz==2017.2
reportlab==3.4.0
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.0.6
